+++ bin/cleanexamples.sh
@@ -30,23 +30,19 @@ rm -rf ./src/*
 touch ./src/.gitkeep
 
 #clean out tests
-mv ./tests/Codeception/Acceptance/ExampleCest.php ./tests/Codeception/Acceptance/ExampleCest.bak
 mv ./tests/Codeception/Acceptance/_bootstrap.php ./tests/Codeception/Acceptance/_bootstrap.bak
 rm -rf ./tests/Codeception/Acceptance/*.php
-mv ./tests/Codeception/Acceptance/ExampleCest.bak ./tests/Codeception/Acceptance/ExampleCest.php
 mv ./tests/Codeception/Acceptance/_bootstrap.bak ./tests/Codeception/Acceptance/_bootstrap.php
 
 mv ./tests/Integration/ExampleTest.php ./tests/ExampleTest.bak
 rm -rf ./tests/Integration/*
 mv ./tests/ExampleTest.bak ./tests/Integration/ExampleTest.php
 
-mv ./tests/Unit/ExampleTest.php ./tests/ExampleTest.bak
 rm -rf ./tests/Unit/*
-mv ./tests/ExampleTest.bak ./tests/Unit/ExampleTest.php
 
 perl -pi\
   -e 's#paths:.*#paths: "Application/views/flow/translations,Application/views/admin"#g;'\
-  ./tests/Codeception/acceptance.suite.yml
+  ./tests/Codeception/Acceptance.suite.yml
 
 echo
 echo 'Cleanup done. Now it is your turn to fill the module with code.'